requirements.txt
requests==2.31.0
pymongo==4.6.1
python-dotenv==1.0.0 
psycopg2==2.9.10